requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
eventlet==0.22.1
Flask==0.12.2
Flask-SocketIO==2.9.6
greenlet==0.4.13
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
netifaces==0.10.6
nltk==3.2.5
Pillow==5.1.0
python-engineio==2.0.4
python-socketio==1.9.0
requests==2.18.4
six==1.11.0
socketio==0.0.7
textblob==0.15.1
urllib3==1.22
Werkzeug==0.14.1
